feat: Anime Auth guard

+++ src/anime/anime.service.ts
@@ -1,10 +1,15 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateAnimeDto } from './dto/create-anime.dto';
 import { UpdateAnimeDto } from './dto/update-anime.dto';
 import { AnimeErrorText } from './anime.enum';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Anime } from './entities/anime.entity';
 import { Repository } from 'typeorm';
+import { User } from '../auth/entities/user.entity';
 
 @Injectable()
 export class AnimeService {
@@ -12,9 +17,18 @@ export class AnimeService {
     @InjectRepository(Anime)
     private animeRepository: Repository<Anime>,
   ) {}
-  async createAnime(createAnimeDto: CreateAnimeDto) {
+  async createAnime(createAnimeDto: CreateAnimeDto, user: User) {
     const { title, author, tag, source } = createAnimeDto;
-    // return await this.animeRepository.save(newAnime);
+
+    const anime = this.animeRepository.create({
+      title,
+      author,
+      tag,
+      source,
+      user,
+    });
+
+    return await this.animeRepository.save(anime);
   }
 
   async getAllAnime() {
@@ -31,21 +45,35 @@ export class AnimeService {
     return anime;
   }
 
-  async updateAnime(id: number, updateAnimeDto: UpdateAnimeDto) {
-    const anime = await this.animeRepository.findOneBy({ id });
+  async updateAnime(id: number, updateAnimeDto: UpdateAnimeDto, user) {
+    const anime = await this.animeRepository.findOne({
+      where: {
+        id,
+      },
+      relations: ['user'],
+    });
+
+    if (anime.user.id !== user.id) {
+      throw new ForbiddenException();
+    }
+
     const newAnime = {
       ...anime,
-      title: updateAnimeDto.title,
-      author: updateAnimeDto.author,
-      tag: updateAnimeDto.tag,
-      source: updateAnimeDto.source,
+      ...updateAnimeDto,
     };
 
     return await this.animeRepository.save(newAnime);
   }
 
-  async removeAnime(id: number) {
-    const anime = await this.animeRepository.findOneBy({ id });
+  async removeAnime(id: number, user: User) {
+    const anime = await this.animeRepository.findOne({
+      where: { id },
+      relations: ['user'],
+    });
+
+    if (user.id !== anime.user.id) {
+      throw new ForbiddenException();
+    }
 
     return await this.animeRepository.remove(anime);
   }
